% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{backtransform_parms}
\alias{backtransform_parms}
\title{Back-transform parameters from the
sampled form to the model and/or prior form}
\usage{
backtransform_parms(parms, rep_df)
}
\description{
Back-transform parameters from the
sampled form to the model and/or prior form
}
