% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npr_estimation_new.R
\name{balance_het}
\alias{balance_het}
\title{Balance the number of partnerships between groups,
The balanced partnership rate is called "corrected_r"}
\usage{
balance_het(df, theta = 0.5)
}
\description{
Balance the number of partnerships between groups,
The balanced partnership rate is called "corrected_r"
}
