% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_force_of_inf.R
\name{calculate_force_of_inf}
\alias{calculate_force_of_inf}
\title{Calculate the force of infection}
\usage{
calculate_force_of_inf(suff_contact_matrix, prev)
}
\arguments{
\item{suff_contact_matrix}{this is the contact matrix multiplied
by the probability of transmission}

\item{prev}{the prevalence vector}
}
\description{
The force of infection is
calculated as the matrix product of the
sufficient contact matrix and the prevalence
}
