% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{define_calib_parms}
\alias{define_calib_parms}
\title{take back-transformed calib results and pass to define_parms}
\usage{
define_calib_parms(i, btparms, so, rep_df, prop_df)
}
\description{
take back-transformed calib results and pass to define_parms
}
