% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_contact_matrix.R
\name{define_contact_matrix}
\alias{define_contact_matrix}
\title{todo: write test}
\usage{
define_contact_matrix(contact_df, parm_indices)
}
\arguments{
\item{contact_df}{data.frame with partnering rates in the \code{corrected_r} column}

\item{parm_indices}{data.frame with the mapping of demographic groups to unique identifiers}
}
\description{
todo: write test
}
