% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_transmission_prob_matrix.R
\name{define_transmission_prob_matrix}
\alias{define_transmission_prob_matrix}
\title{define transmission probabilities between sexes}
\usage{
define_transmission_prob_matrix(betaMM, betaMW, betaWW, demo_indices)
}
\arguments{
\item{betaMM}{prob. between men}

\item{betaMW}{prob. between men and women (assumed to be the same in both directions)}

\item{betaWW}{prob. between women}

\item{demo_indices}{data frame mapping demographic groups to unique identifiers}
}
\description{
transmission probabilities are by partnership
}
