% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imis_functions.R
\name{estimate_inf_duration}
\alias{estimate_inf_duration}
\title{Use data to estimate gamma distributions for the infection duration
Returns parameters for gamma distributions, male and female respectively}
\usage{
estimate_inf_duration()
}
\description{
Use data to estimate gamma distributions for the infection duration
Returns parameters for gamma distributions, male and female respectively
}
