% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{get_demo_index}
\alias{get_demo_index}
\title{Get the index for a demographic group}
\usage{
get_demo_index(qsex, qsexid, qsexact, demo_indices)
}
\description{
Get the index for a given demographic group.
This is the demo index (see \code{\link{define_index_df}}),
not the model index.
}
