% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{group_prevalence_to_gender}
\alias{group_prevalence_to_gender}
\title{Takes formatted prevalence to gender-specific prevalence}
\usage{
group_prevalence_to_gender(fprev)
}
\description{
Takes formatted prevalence to gender-specific prevalence
}
