% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{likelihood_generator}
\alias{likelihood_generator}
\title{Used to define a likelihood function either for het or so mixing structures}
\usage{
likelihood_generator(so, rep_df)
}
\arguments{
\item{so}{sex orientation: either "het" or "msid"}

\item{rep_df}{reported partnerships dataframe}
}
\description{
Used to define a likelihood function either for het or so mixing structures
}
