% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{load_constants}
\alias{load_constants}
\title{Load constants}
\usage{
load_constants(ids)
}
\arguments{
\item{ids}{either msid, het, or all}
}
\description{
These are used throughout the code to
keep track of compartments
}
