% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npr_estimation_new.R
\name{make_offered_df}
\alias{make_offered_df}
\title{This calculates the proportion of partnerships
"offered to" sex A that come from people with
sex r_sex, sex activity r_sexact, and sexid r_sexid}
\usage{
make_offered_df(df)
}
\description{
This calculates the proportion of partnerships
"offered to" sex A that come from people with
sex r_sex, sex activity r_sexact, and sexid r_sexid
}
