% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_parameters.R
\name{make_parm_gender_vecs}
\alias{make_parm_gender_vecs}
\title{for parameters with a male value and a female value
takes the value and turns into a vector of male and female values}
\usage{
make_parm_gender_vecs(mval, wval, demo_indices)
}
\description{
for parameters with a male value and a female value
takes the value and turns into a vector of male and female values
}
