% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_function.R
\name{model_function}
\alias{model_function}
\title{Contains the differential equations that describe the
HPV transmission model}
\usage{
model_function(times, x, parms, vaccination)
}
\description{
Contains the differential equations that describe the
HPV transmission model
}
