% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{prior}
\alias{prior}
\title{prior density of parms.
takes parameters directly from sample.prior()}
\usage{
prior(parms)
}
\description{
prior density of parms.
takes parameters directly from sample.prior()
}
