% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{prior_generator}
\alias{prior_generator}
\title{prior density of parms.
takes parameters directly from sample.prior()}
\usage{
prior_generator(rep_df, so)
}
\description{
prior density of parms.
takes parameters directly from sample.prior()
}
