% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa.R
\name{psa_analysis}
\alias{psa_analysis}
\title{run analysis once with sampled parameters}
\usage{
psa_analysis(so, psa_parms, phi_base, phi_comp, poc = c("prev", "cases"))
}
\arguments{
\item{imis_parameters}{assumed to be back-transformed}
}
\description{
run analysis once with sampled parameters
}
