% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{replace_filtered_nprs}
\alias{replace_filtered_nprs}
\title{take parameters (sampled from prior) and
replace the 0-nprs that were removed earlier}
\usage{
replace_filtered_nprs(filt_nprs, so)
}
\description{
take parameters (sampled from prior) and
replace the 0-nprs that were removed earlier
}
