% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{run_model_with_btparms}
\alias{run_model_with_btparms}
\title{Run the model with the ith set of bback-transformed parameters}
\usage{
run_model_with_btparms(i, btparms, so, pop_dist, contact)
}
\description{
Called from the likelihood function.
}
