% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_to_steady.R
\name{run_to_steady}
\alias{run_to_steady}
\title{Run the model to steady state, based on parameters and vaccination}
\usage{
run_to_steady(parms, vaccination)
}
\description{
Run the model to steady state, based on parameters and vaccination
}
