% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{sample_prior_generator}
\alias{sample_prior_generator}
\title{Sample from the prior distributions.}
\usage{
sample_prior_generator(rep_df)
}
\description{
Called from IMIS.
}
\details{
In this function, the prior samples are transformed to be unbounded, which
makes a better search space for IMIS. Then, the parameters are "back-transformed"
before being sent to the model function.
}
