% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_contact_matrix.R
\name{define_contact_matrix}
\alias{define_contact_matrix}
\title{define partnering rates between demographic groups}
\usage{
define_contact_matrix(contact_df, demo_indices)
}
\arguments{
\item{contact_df}{data.frame with partnering rates in the \code{} column}

\item{demo_indices}{data.frame with the mapping of demographic groups to unique identifiers}
}
\description{
define partnering rates between demographic groups
}
