% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_constants}
\alias{load_constants}
\title{These are used throughout the code to
keep track of compartments and such}
\usage{
load_constants(ids)
}
\description{
These are used throughout the code to
keep track of compartments and such
}
